Project Finished
Ödev tamamlandı

+++ Hafta-1-Algoritma-Soruları-Can Önal/src/soru4/Soru4.java
@@ -3,8 +3,8 @@ package soru4;
 public class Soru4 {
 
 	public static void main(String[] args) {
-		for (int i = 1; i < 10; i++) {
-			for (int j = 0; j < i; j++) {
+		for (byte i = 1; i < 10; i++) {
+			for (byte j = 0; j < i; j++) {
 				System.out.print("*");
 			}
 			System.out.println();
